requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==4.0.3
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.2
cryptography==46.0.6
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.1
fastf1==3.8.2
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-experimental==0.4.1
langchain-groq==1.1.2
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.18.1
numpy==2.2.6
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas==2.3.3
peewee==4.0.3
pillow==12.1.1
platformdirs==4.9.4
primp==1.2.1
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
pyparsing==3.3.2
pyrate-limiter==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.33.1
requests-cache==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
scipy==1.15.3
sec-edgar-downloader==5.1.0
signalrcore==1.0.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
streamlit==1.55.0
tenacity==9.1.4
timple==0.1.8
toml==0.10.2
tornado==6.5.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zstandard==0.25.0
